Separate FastAPI and Starlette utilities
br3ndonland/inboard@81d5a90
br3ndonland/inboard@a3ffa71
br3ndonland/inboard@823408d
br3ndonland/inboard@c40a020
br3ndonland/inboard@9c79964

This commit will separate the utilities for FastAPI and Starlette into
separate modules, so that the Starlette application can run if FastAPI
is not installed.
