+++ inboard/app/main_fastapi.py
@@ -5,7 +5,7 @@ from typing import Dict
 from fastapi import Depends, FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from inboard.app.utilities import basic_auth
+from inboard.app.utilities_fastapi import basic_auth
 
 server = "Uvicorn" if bool(os.getenv("WITH_RELOAD")) else "Uvicorn, Gunicorn"
 version = f"{sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}"
